refactored code to better support jupyter lab; added graph.py and predict.py
