+++ .gitignore
@@ -113,3 +113,4 @@ dmypy.json
 
 # other
 todo.txt
+.DS_Store
